Eliminate `TextConversion`
