+++ lib/spectacular/src/core/spectacular.Showable.scala
@@ -36,14 +36,46 @@ import anticipation.*
 import denominative.*
 import digression.*
 import fulminate.*
+import inimitable.*
+import proscenium.*
+import rudiments.*
 import vacuous.*
 
 import language.experimental.captureChecking
 
-trait Showable extends TextConversion:
+trait Showable:
   type Self
+  def text(value: Self): Text
 
 object Showable:
+  given showable: [value: Textualizer] => value is Showable = value.textual(_)
+
+  given text: [text <: Text] => text is Showable = identity(_)
+  given string: String is Showable = _.tt
+  given char: Char is Showable = char => char.toString.tt
+  given long: Long is Showable = long => long.toString.tt
+  given int: Int is Showable = int => int.toString.tt
+  given short: Short is Showable = short => short.toString.tt
+  given byte: Byte is Showable = byte => byte.toString.tt
+  given message: Message is Showable = _.text
+  given double: (decimalizer: DecimalConverter) => Double is Showable = decimalizer.decimalize(_)
+  given boolean: (booleanStyle: BooleanStyle) => Boolean is Showable = booleanStyle(_)
+  given option: [value: Showable] => Option[value] is Showable = _.fold("none".tt)(value.text(_))
+  given uid: Uuid is Showable = _.text
+  given memory: Memory is Showable = _.text
+  given enumeration: [enumeration <: reflect.Enum] => enumeration is Showable = _.toString.tt
+
+  given set: [element: Showable] => Set[element] is Showable =
+    _.map(_.show).mkString("{", ", ", "}").tt
+
+  given list: [element: Showable] => List[element] is Showable =
+    _.map(_.show).mkString("[", ", ", "]").tt
+
+  given trie: [element: Showable] => Trie[element] is Showable =
+    _.map(_.show).mkString("[ ", " ", " ]").tt
+
+  given none: None.type is Showable = none => "none".tt
+
   given specializable: Specializable is Showable = value =>
     value.getClass.nn.getName.nn.split("\\.").nn.last.nn.dropRight(1).toLowerCase.nn.tt
